Python 3 and more!
* Port to Python 3
* Use `fInitializeProduct.py`
* Use `mExports.py`
* dxProductDetails version 2021-07-02 11:16
* Standardize `Tests.cmd`
* Improve `Tests\fTestDependencies.py`
* Standardize variable/function/method naming convention
* replace `oConsole.fPrint` with `oConsole.fOutput`
* Use standardized `fatsArgumentLowerNameAndValue` to process arguments.
* Use standardized `fPrintLicenseInformation` and `fPrintVersionInformation`.
* Standardize `mColors.py`

* Make sure ALL pipes are closed when a `cConsoleProcess` terminates.
* `cPipe` now has `.bClosed`, `bClosedForReading` and `bClosedForWriting`.
* Various bug fixes and improvements

+++ cJobObject.py
@@ -86,21 +86,21 @@ class cJobObject(object):
   
   def fSetMaxProcessMemoryUse(oSelf, uMemoryUseInBytes):
     oExtendedLimitInformation = oSelf.__foQueryExtendedLimitInformation();
-    oExtendedLimitInformation.ProcessMemoryLimit = long(uMemoryUseInBytes);
+    oExtendedLimitInformation.ProcessMemoryLimit = int(uMemoryUseInBytes);
     oExtendedLimitInformation.BasicLimitInformation.LimitFlags |= JOB_OBJECT_LIMIT_PROCESS_MEMORY;
     oSelf.__fSetExtendedLimitInformation(oExtendedLimitInformation);
 
   def fSetMaxTotalMemoryUse(oSelf, uMemoryUseInBytes):
     oExtendedLimitInformation = oSelf.__foQueryExtendedLimitInformation();
-    oExtendedLimitInformation.JobMemoryLimit = long(uMemoryUseInBytes);
+    oExtendedLimitInformation.JobMemoryLimit = int(uMemoryUseInBytes);
     oExtendedLimitInformation.BasicLimitInformation.LimitFlags |= JOB_OBJECT_LIMIT_JOB_MEMORY;
     oSelf.__fSetExtendedLimitInformation(oExtendedLimitInformation);
   
   def fuGetMaxProcessMemoryUse(oSelf):
     oExtendedLimitInformation = oSelf.__foQueryExtendedLimitInformation();
-    return long(oExtendedLimitInformation.PeakProcessMemoryUsed);
+    return int(oExtendedLimitInformation.PeakProcessMemoryUsed);
   
   def fuGetMaxTotalMemoryUse(oSelf):
     oExtendedLimitInformation = oSelf.__foQueryExtendedLimitInformation();
-    return long(oExtendedLimitInformation.PeakJobMemoryUsed);
+    return int(oExtendedLimitInformation.PeakJobMemoryUsed);
